Searchable Encryption Final Python Version
Example: Assume that Alice want to redirect the mail to Bob containing the term “office” as she is away in a holiday. Now how would the service provider know which mail to send to Bob if it is not able to read the data ? Again it is a security concern to transmit the data with out end to end encryption. A solution of this problem is to use Searchable Encryption.